refactor current solution
