--- packages/react/src/components/utils/chat-floating-button/chat-floating-button.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import SpeedDialAction from 'devextreme-react/speed-dial-action';
-
-type ChatFloatingButtonProps = {
-  onClick: () => void;
-};
-
-export const ChatFloatingButton = ({ onClick }: ChatFloatingButtonProps) => {
-  return <SpeedDialAction icon='sparkle' label='AI Inside' onClick={onClick} />;
-};
